requirements.txt
requests
fake_useragent
google-api-python-client
google.oauth2
googleapiclient.discovery